# docs/api/swagger.yaml
basePath: /api/v1
definitions:
  big.Int:
    type: object
  fmis-api_internal_models.BatchStatusType:
    enum:
    - ACTIVE
    - DEPLETED
    - QUARANTINED
    - EXPIRED
    - RESERVED
    type: string
    x-enum-varnames:
    - BatchStatusTypeActive
    - BatchStatusTypeDepleted
    - BatchStatusTypeQuarantined
    - BatchStatusTypeExpired
    - BatchStatusTypeReserved
...
  fmis-api_internal_schemas.CreateBatchRequest:
    properties:
      batch_number:
        maxLength: 100
        type: string
      expiration_date:
        type: string
      product_id:
        type: string
      quantity:
        type: string
    required:
    - batch_number
    - product_id
    - quantity
    type: object
...
  fmis-api_internal_schemas.InventoryBatch:
    properties:
      batch_number:
        type: string
      created_at:
        type: string
      expiration_date:
        $ref: '#/definitions/pgtype.Timestamptz'
      id:
        type: string
      product_id:
        type: string
      quantity_current:
        $ref: '#/definitions/pgtype.Numeric'
      quantity_initial:
        $ref: '#/definitions/pgtype.Numeric'
      status:
        $ref: '#/definitions/fmis-api_internal_models.BatchStatusType'
      updated_at:
        type: string
    type: object
  fmis-api_internal_schemas.LoginRequest:
    properties:
      identifier:
        maxLength: 255
        type: string
      password:
        type: string
    required:
    - identifier
    - password
    type: object
  fmis-api_internal_schemas.LogoutRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  fmis-api_internal_schemas.MeResponse:
    properties:
      email:
        type: string
      id:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
...
  fmis-api_internal_schemas.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  fmis-api_internal_schemas.RegisterRequest:
    properties:
      email:
        maxLength: 100
        minLength: 3
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      username:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  fmis-api_internal_schemas.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
...
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    format: int32
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
  pgtype.Numeric:
    properties:
      exp:
        format: int32
        type: integer
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      int:
        $ref: '#/definitions/big.Int'
      naN:
        type: boolean
      valid:
        type: boolean
    type: object
...
  pgtype.Timestamptz:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        type: string
      valid:
        type: boolean
    type: object
info:
  contact: {}
  description: Food Manufacturing Inventory System REST API.
  title: FMIS API
  version: 1.0.0
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Verifies credentials by username or email and returns a token pair.
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/fmis-api_internal_schemas.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fmis-api_internal_schemas.TokenResponse'
        "400":
          description: Invalid body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Log in
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revokes a refresh token, ending the session. The access token itself
        remains valid until it expires.
      parameters:
      - description: Refresh token to revoke
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/fmis-api_internal_schemas.LogoutRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "400":
          description: Invalid body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid refresh token
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Log out
      tags:
      - auth
  /auth/me:
    get:
      description: Returns the authenticated user's profile (id, username, email,
        role).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fmis-api_internal_schemas.MeResponse'
        "401":
          description: Missing or invalid token
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchanges a valid refresh token for a new access token + refresh
        token pair.
      parameters:
      - description: Refresh token payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/fmis-api_internal_schemas.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fmis-api_internal_schemas.TokenResponse'
        "400":
          description: Invalid body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid or expired refresh token
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh an access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a user with the default VIEWER role and returns a token
        pair. Requires a unique username and email.
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/fmis-api_internal_schemas.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/fmis-api_internal_schemas.TokenResponse'
        "400":
          description: Invalid body
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Username or email already taken
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /batches:
    post:
      consumes:
      - application/json
      description: Registers an incoming lot of a product. ADMIN or OPERATOR role
        required.
      parameters:
      - description: Batch payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/fmis-api_internal_schemas.CreateBatchRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/fmis-api_internal_schemas.InventoryBatch'
        "400":
          description: Invalid body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Missing or invalid token
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Insufficient role
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Expiration date required for this product
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Receive a batch
      tags:
      - batches
  /batches/{batch_id}/quarantine:
    patch:
      description: Marks an ACTIVE batch as QUARANTINED so it is excluded from FEFO
        consumption. ADMIN or OPERATOR role required.
      parameters:
      - description: Batch UUID
        in: path
        name: batch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fmis-api_internal_schemas.InventoryBatch'
        "400":
          description: Malformed batch id
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Missing or invalid token
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Insufficient role
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Batch not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Batch not in ACTIVE state
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Quarantine a batch
      tags:
      - batches
  /batches/product/{product_id}:
    get:
      description: Returns all non-depleted, non-quarantined batches for a product,
        ordered FEFO (soonest expiration first).
      parameters:
      - description: Product UUID
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/fmis-api_internal_schemas.InventoryBatch'
            type: array
        "400":
          description: Malformed product id
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Missing or invalid token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List active batches for a product
      tags:
      - batches
...
securityDefinitions:
  BearerAuth:
    description: 'Pass the JWT access token as `Authorization: Bearer <token>`.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"